+++ tests/Functional/DoctrineOrm/Entities/Product.php
@@ -48,4 +48,8 @@ class Product
 
     #[ORM\Embedded(class: Total::class, columnPrefix: 'total_')]
     private Total $total;
+
+    #[ORM\ManyToOne(targetEntity: ProductGroup::class)]
+    #[ORM\JoinColumn(name: 'group_id', referencedColumnName: 'id')]
+    private ProductGroup $group;
 }